requirements.txt

requests
python-dotenv
schedule
twilio
gTTS
pydub
GitPython
Flask
